add generic openai-compatible API, add credential template for that
(cherry picked from commit 0e1e0d36906e77e09abddfa5a6464ba0f2624079)
